I separated 'regular' genetic algorithm code from the parallel code, creating new files and directories
